pyproject.toml
[tool.poetry]
name = "user_database_tg"
version = "0.1.0"
description = ""
authors = ["taimaskhanov11 <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
loguru = "^0.6.0"
tortoise-orm = { extras = ["asyncpg"], version = "^0.18.1" }
aiogram = "^2.19"
pyQiwiP2P = "^1.0.1"
httpx = "^0.22.0"
python-dotenv = "^0.20.0"
environs = "^9.5.0"
pydantic = "^1.9.0"
mtranslate = "^1.8"
aiohttp = "^3.8.1"
socid-extractor = "^0.0.23"
geopy = "^2.2.0"
selenium-wire = "^4.6.3"
selenium = "^4.1.3"
ImageHash = "^4.2.1"
Pillow = "^9.0.1"
python-dateutil = "^2.8.2"
colorama = "^0.4.4"
beautifultable = "^1.0.1"
termcolor = "^1.1.0"
webdriver-manager = "^3.5.4"
wayback = "^0.3.2"
bs4 = "^0.0.1"
packaging = "^21.3"
aiofiles = "^0.8.0"
APScheduler = "^3.9.1"
fastapi = "^0.76.0"
uvicorn = "^0.17.6"
slowapi = "^0.1.5"
hashID = "^3.1.4"


[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
peewee = "^3.14.9"
isort = "^5.10.1"
black = "^22.1.0"
pyftpdlib = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
